android-6.0.0_r1

+++ external/apache-harmony/jdwp/src/test/java/org/apache/harmony/jpda/tests/jdwp/DebuggerOnDemand/OnthrowLaunchDebugger002.java
@@ -61,7 +61,7 @@ public class OnthrowLaunchDebugger002 extends LaunchedDebugger {
         }
 
         logWriter.println("**> Set breakpoint at the beginning of " + breakpointMethodName);
-        long requestID = 0;
+        int requestID = 0;
         try {
             requestID = debuggeeWrapper.vmMirror.setBreakpointAtMethodBegin(
                         refTypeID, breakpointMethodName);
